documentation-system/requirements.txt
Flask
flask-cors
pdf2image
pytesseract
transformers
torch
spacy
python-dotenv